Move functions using io-ts to a separate file
