+++ extensions/ql-vscode/src/remote-queries/gh-api/gh-api-client.ts
@@ -1,8 +1,6 @@
-import * as t from "io-ts";
-import { fold, isLeft } from "fp-ts/Either";
-import { pipe } from "fp-ts/function";
 import { Credentials } from "../../authentication";
 import { OctokitResponse } from "@octokit/types/dist-types";
+import { validateApiResponse } from "../../pure/io-ts";
 import { RemoteQueriesSubmission } from "../shared/remote-queries";
 import { VariantAnalysisSubmission } from "../shared/variant-analysis";
 import {
@@ -15,54 +13,6 @@ import {
   RemoteQueriesResponse,
   RemoteQueriesSubmissionRequest,
 } from "./remote-queries";
-
-function stringify(v: any): string {
-  if (typeof v === "function") {
-    return t.getFunctionName(v);
-  }
-  if (typeof v === "number" && !isFinite(v)) {
-    if (isNaN(v)) {
-      return "NaN";
-    }
-    return v > 0 ? "Infinity" : "-Infinity";
-  }
-  return JSON.stringify(v);
-}
-
-function getContextPath(context: t.Context): string {
-  return context.map(({ key }) => key).join(".");
-}
-
-function getMessage(e: t.ValidationError): string {
-  return e.message !== undefined
-    ? e.message
-    : `Invalid value ${stringify(e.value)} supplied to ${getContextPath(
-        e.context,
-      )}`;
-}
-
-const getErrors = <A>(v: t.Validation<A>): string[] => {
-  return pipe(
-    v,
-    fold(
-      (errors) => errors.map(getMessage),
-      () => ["no errors"],
-    ),
-  );
-};
-
-function validateApiResponse<T extends t.Any>(
-  data: unknown,
-  type: T,
-): t.TypeOf<T> {
-  const result = type.decode(data);
-  if (isLeft(result)) {
-    throw new Error(
-      `Invalid response from GitHub API: ${getErrors(result).join(", ")}`,
-    );
-  }
-  return result.right;
-}
 
 export async function submitVariantAnalysis(
   credentials: Credentials,
